Fix the error of multi link ide test case
In x4x4x4x4 case, KCBar is shared between rootport_1 and rootport_2. So
the previous logic which is to find a free stream_x is not correct.

The updated logic to find a free stream_x and Link/Selective IDE
Register block is simplified that it walks thru Link/Selective IDE
Register blocks and Stream_x in KCBar to find a free slot. It will not
check the stream_id.

Signed-off-by: Min Xu <min.m.xu@intel.com>