Set Addr Assoc Register
Roorport Addr Assoc Register shall be set based on device card.
If the device card is 32bit, then Rootport Addr Assoc Register
shall be populated with the value from device's Memory Lit/Base.
If the device is 64bit, it shall be populated with the value from
device's Prefetchable Memory/Base.

As to the Device Addr Assoc Register the addr assoc2 is set as fixed
value (0x0).

Signed-off-by: Min Xu <min.m.xu@intel.com>